clean up dashboard
