+++ src/views/dashboard/Dashboard.tsx
@@ -1,12 +1,12 @@
 import React, { ReactElement } from "react";
 import { LessonsWidget } from "./LessonsWidget";
 // import { SetsWidget } from "./SetsWidget";
-import { ActivityWidget } from "./ActivityWidget";
 import { GettingStartedWidget } from "./GettingStartedWidget";
 import { useUserStateContext } from "../../providers/UserStateProvider";
 import { OPEN_BETA_ID } from "../../state/reducers/groupId";
 import { MinigameWidget } from "./MinigameWidget";
 import { useAnalyticsPageName } from "../../firebase/hooks";
+import { GetHelpWidget } from "./GetHelpWidget";
 
 export function Dashboard(): ReactElement {
   const {
@@ -18,8 +18,8 @@ export function Dashboard(): ReactElement {
       {/** Only show getting started for folks in open beta */}
       {groupId === OPEN_BETA_ID && <GettingStartedWidget />}
       <MinigameWidget />
-      <ActivityWidget />
       <LessonsWidget />
+      <GetHelpWidget />
     </div>
   );
 }
